+++ yaqc_qtpy/property_items/_string.py
@@ -1,0 +1,29 @@
+__all__ = ["String"]
+
+
+import time
+from functools import partial
+
+import qtypes
+
+
+def value_updated(value, item):
+    item.set({"value": item.get()})
+
+
+def set_daemon(value, property):
+    property(value["value"])
+
+
+def String(key, property, qclient):
+    # disabled
+    if property._property.get("setter", None):
+        disabled = False
+    else:
+        disabled = True
+    # make item
+    item = qtypes.String(disabled=disabled, label=key)
+    # signals and slots
+    property.updated.connect(partial(value_updated, item=item, units=default_units.result))
+    item.edited.connect(partial(set_daemon, property=property))
+    return item